Work On RF
- DMA: DMA configurations for RFRX are confirmed to work.
--> Will need to consider how to handle for TX if we want to use
DMA. RF uses the same interrupt vector for TXRX.
SUGGESTON: For RFRX, will probably be best to arm channel 2 for
data transfers and creating DMAREQ to allow for data to be sent
for TX. increment source, but not destination. When interrupts,
it should check for the appropriate source channel.

RFRX: Have gotten RFRX to receive and return `Packet Received`
through UART.
--> Currently have a bunch of test uart sends scattered around to
validate RFRX path.
--> In SmartRF Studio, the packet that was used was `ACKNOWLDGE`
but with the length byte removed as it is applied through software.
--> RX didn't look to overflow at all with the usage of DMA, unlike
before.
NOTE: For register PKTCTRL1, when Append status is included the RX
path seems to not function correctly.

- RFTX: Still a work in progress, but have had better success with the
continuous RX graph, with nicer and more consistent power spikes.
--> SmartRF Studio can pick up the incoming packets with Pkt RX,
but tends to always fail CRC check.
--> Length of time between each packet being sent may not be long
enough. Probably will need to add delay closer to 700ms.
--> Will need to consider how to handle for DMA if used.
--> Need to double check if the AppendStatus is important for TX.

- Packet Handler: Minor changes to allow for better testing of path
coverage for RFRX. With a packet sent for `ACKNOWLEDGE` we enter
the switch case for the corresponding command. Still need to check
for others.

- UART: UART is working as expected, and now being implemented
as a test tool for path coverage.