MAJOR CHANGES
- SDCC: We have moved to using Small Device C Compiler over Keil because of
the limits for compiling code on the free version of Keil. This change required a lot
of messing around to get it to work, but finally I have it in a stable state that I believe
we can continue working from.
--> Remade the file organization structure to work better with Makefile and SDCC
compiler handling.
--> you will need to make  sure that you have the following directories in your project
folder: src, include, obj, and bin. Use command `mkdir -p folder_name` to create these.
--> I have also inclduded the manual for SDCC if you wish to go over any of the
content for yourself.

-Makefile: Used to compile and build the code base together rather than individually
doing each step of the process.
--> Look at file for the `SRC_DIR`, `OBJ_DIR`, and `BIN_DIR` to make sure that they are
the same as the ones you have created. If not update them or create according to the
command stated in SDCC.

- *.c files: No major changes besides being added to the src folder all together.

- *.h files: No major changes besides being added to the include folder all together
plus 2 new file additions.
--> compiler.h: This file is used to define the compiler that is being used and how to
handle defining the functions for sfr and sbit.
--> 2510fx.h: same as the old 2510.h file, but with the addition of the new sfr and sbit
definitions to conform with SDCC.

SDCC_Setup.txt: General notes and simple guide on how to compile and build code
for flashing the chip. Not super ingrained but should be able to give general starting block on how to get everything going.

NOTE: When working with the Makefile we must explicitly define the ranges of
memory that the CC chip can work with. Without it, we will get issues with reading
memory correctly, and potenitally get stuck in infinite loops.